+++ diesel/src/sqlite/mod.rs
@@ -24,12 +24,19 @@ pub trait SqliteAggregateFunction<Args>: Default {
     /// The result type of the SQLite aggregate function
     type Output;
 
-    /// The `step()` method is called once for every record of the query
+    /// The `step()` method is called once for every record of the query.
+    ///
+    /// This is called through a C FFI, as such panics do not propagate to the caller. Panics are
+    /// caught and cause a return with an error value. The implementation must still ensure that
+    /// state remains in a valid state (refer to [`std::panic::UnwindSafe`] for a bit more detail).
     fn step(&mut self, args: Args);
 
     /// After the last row has been processed, the `finalize()` method is
     /// called to compute the result of the aggregate function. If no rows
     /// were processed `aggregator` will be `None` and `finalize()` can be
-    /// used to specify a default result
+    /// used to specify a default result.
+    ///
+    /// This is called through a C FFI, as such panics do not propagate to the caller. Panics are
+    /// caught and cause a return with an error value.
     fn finalize(aggregator: Option<Self>) -> Self::Output;
 }