+++ diesel_derives/src/lib.rs
@@ -1011,6 +1011,14 @@ pub fn derive_valid_grouping(input: TokenStream) -> TokenStream {
 /// # }
 /// ```
 ///
+/// ## Panics
+///
+/// If an implementation of the custom function panics and unwinding is enabled, the panic is
+/// caught and the function returns to libsqlite with an error. It cannot propagate the panics due
+/// to the FFI bounary.
+///
+/// This is is the same for [custom aggregate functions](#custom-aggregate-functions).
+///
 /// ## Custom Aggregate Functions
 ///
 /// Custom aggregate functions can be created in SQLite by adding an `#[aggregate]`